--- app/assets/stylesheets/departments/maintenance.scss
@@ -1,28 +1,0 @@
-
-div.maintenance-header{
-  width: 100vw;
-  height: 70px;
-  background-color: red;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  opacity: 0.7;
-  margin-bottom: 20px;
-}
-
-.maintenance-title{
-  color: aliceblue;
-  font-size: 28px;
-}
-
-.maintenance-links{
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
-
-div.maintenance-header div ul.maintenance-links li{
-  padding: 5px 30px;
-  text-decoration: none;
-}
