+++ app/assets/stylesheets/user_departments_relations.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the user_departments_relations controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
