+++ app/assets/stylesheets/user_departments_relations/maintenance.scss
@@ -1,0 +1,63 @@
+
+// indexページでのCSS
+div.maintenance-header{
+  width: 100vw;
+  height: 75px;
+  background-color: salmon;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 20px;
+}
+
+.maintenance-title{
+  color: aliceblue;
+  font-size: 28px;
+}
+
+.maintenance-links{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+}
+
+div.maintenance-header div ul.maintenance-links li{
+  padding: 5px 35px;
+  text-decoration: none;
+  text-align: center;
+}
+
+
+// new-relationでのCSS
+div.new-relation{
+  width: 40vw;
+  height: calc(100vh -75px);
+  font-size: 30px;
+  margin-left: 30vw;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+div.new-relation div.new-relation-field{
+  margin: 10px 0;
+  padding: 5px 0;
+  padding-left: 20px;
+  background-color: palevioletred;
+}
+
+div.new-relation div.new-relation-field .new-relation-select-box{
+  width: 90%;
+  height: auto;
+}
+
+.new-relation-checkbox{
+  padding-left: 30px;
+  transform:scale(3);
+}
+
+div.new-relation div.new-relation-field div.submit{
+  display: flex;
+  justify-content: center;
+}
