+++ app/helpers/user_departments_relations_helper.rb
@@ -1,0 +1,2 @@
+module UserDepartmentsRelationsHelper
+end
