--- app/views/departments/_header.html.erb
@@ -1,11 +1,0 @@
-<div class="maintenance-header">
-  <p class="maintenance-title">
-    社員部署メンテナンスページ
-  </p>
-  <div>
-    <ul class="maintenance-links">
-      <li><%= link_to "トップ", root_path %></li>
-      <li><%= link_to "新規 部署割当", new_department_path %></li>
-    </ul>
-  </div>
-</div>
