+++ spec/requests/user_departments_relations_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "UserDepartmentsRelations", type: :request do
+
+end
